+++ workflows/business_case_workflow/workflow.py
@@ -7,7 +7,7 @@ from models.business_case import BusinessCaseOutput, BusinessCaseResponse
 from typing import TypedDict, Optional, List
 from workflows.nodes import get_chat_history, get_content_file
 from connect_model import get_model_client, set_request_model_config, reset_request_model_config, MODEL
-from ..utils import extractor
+from utils import extractor
 from response import success_response, error_response
 
 class BusinessCaseState(TypedDict):
