+++ workflows/compliance_workflow/workflow.py
@@ -8,7 +8,7 @@ from models.compliance import ComplianceOutput, ComplianceResponse
 from typing import TypedDict, Optional, List
 from workflows.nodes import get_chat_history, get_content_file
 from connect_model import get_model_client, set_request_model_config, reset_request_model_config, MODEL
-from ..utils import extractor
+from utils import extractor
 from response import success_response, error_response
 class ComplianceState(TypedDict):
     user_message: str
