+++ workflows/lld_db_workflow/workflow.py
@@ -10,7 +10,7 @@ import re
 from workflows.nodes import get_chat_history, get_content_file
 from connect_model import get_model_client, set_request_model_config, reset_request_model_config, MODEL
 # from models.lld_db import LLDDBResponse, LLDDBOutput
-from ..utils import extractor
+from utils import extractor
 from response import success_response, error_response
 
 logger = logging.getLogger(__name__)
